+++ src/listpick/utils/utils.py
@@ -12,6 +12,7 @@ from math import log10
 import subprocess
 import tempfile
 import os
+from typing import Tuple, Dict
 
 def truncate_to_display_width(text: str, max_column_width: int, centre=False) -> str:
     """ 
@@ -137,6 +138,26 @@ def get_selected_indices(selections: dict[int, bool]) -> list[int]:
     # selected_indices = [items[i] for i, selected in selections.values() if selected]
     selected_indices = [i for i, selected in selections.items() if selected]
     return selected_indices
+
+def get_selected_cells(cell_selections: Dict[Tuple[int, int], bool]) -> list[Tuple[int, int]]:
+    """ {(0,1): True, (9,1): True} """
+    selected_cells = [i for i, selected in cell_selections.items() if selected]
+    return selected_cells
+
+def get_selected_cells_by_row(cell_selections: dict[tuple[int, int], bool]) -> dict[int, list[int]]:
+    """ {0: [1,2], 9: [1] }"""
+    
+    d = {}
+    try:
+        for tup in cell_selections.keys():
+            if cell_selections[tup]:
+                if tup[0] in d:
+                    d[tup[0]].append(tup[1])
+                else:
+                    d[tup[0]] = [tup[1]]
+    except:
+        pass
+    return d
 
 def get_selected_values(items: list[list[str]], selections: dict[int, bool]) -> list[list[str]]:
     """ Return a list of rows based on wich are True in the selections dictionary. """